free descriptors
* can mark descriptors as free to be reused from free list
* allocation spot for descriptor is now thread safe d3d12

+++ QhenkiX/qhenkiX/graphics/d3d12/d3d12_descriptor_heap.h
@@ -18,13 +18,14 @@ namespace qhenki::gfx
 		std::mutex m_mutex_;
 
 		size_t m_pointer_ = 0;
-		std::vector<unsigned> m_free_list_;
+		std::vector<UINT64> m_free_list_;
 
 	public:
 		const D3D12_DESCRIPTOR_HEAP_DESC& desc = m_desc_;
 		const UINT& descriptor_size = m_descriptor_size_;
 		bool create(ID3D12Device* device, const D3D12_DESCRIPTOR_HEAP_DESC& desc);
 
+		// Thread safe
 		bool allocate(UINT64* alloc_offset);
 		void deallocate(UINT64* alloc_offset);
